refactor offer page with a offer form
