+++ src/components/OfferForm.js
@@ -1,0 +1,381 @@
+import get from 'lodash.get'
+import React, { Component } from 'react'
+import { connect } from 'react-redux'
+import { NavLink } from 'react-router-dom'
+
+import MediationManager from './MediationManager'
+import OccurenceManager from './OccurenceManager'
+import Icon from './layout/Icon'
+import FormField from './layout/FormField'
+import Label from './layout/Label'
+import { mergeForm } from '../reducers/form'
+import { closeModal, showModal } from '../reducers/modal'
+import selectOffererOptions from '../selectors/offererOptions'
+import selectSelectedOffererId from '../selectors/selectedOffererId'
+import selectSelectedVenueId from '../selectors/selectedVenueId'
+import selectSelectedVenues from '../selectors/selectedVenues'
+import selectVenueOptions from '../selectors/venueOptions'
+import { pluralize } from '../utils/string'
+
+class OfferForm extends Component {
+
+  handleMergeForm = () => {
+    const {
+      mergeForm,
+      occasionId,
+      selectedVenueId,
+      uniqueVenue
+    } = this.props
+    uniqueVenue && mergeForm(
+      'occasions',
+      occasionId,
+      'venueId', selectedVenueId
+    )
+  }
+
+  handleShowOccurencesModal = () => {
+    const {
+      currentOccasion,
+      match: { params: { modalType } },
+      showModal
+    } = this.props
+
+    if (modalType !== 'dates') {
+      return
+    }
+
+    showModal(<OccurenceManager occasion={currentOccasion} />)
+
+  }
+
+  handleSuccessData = (state, action) => {
+    const {
+      data,
+      method
+    } = action
+    const {
+      closeModal,
+      history,
+      offerForm,
+      showModal,
+      showNotification
+    } = this.props
+    const {
+      isEventType
+    } = (offerForm || {})
+
+    // PATCH
+    if (method === 'PATCH') {
+      history.push('/offres')
+      showNotification({
+        text: 'Votre offre a bien été enregistrée',
+        type: 'success'
+      })
+      return
+    }
+
+    // POST
+    if (method === 'POST') {
+      // switch to the path with the new created id
+      const routePath = `/offres/${isEventType ? 'evenements' : 'objets'}/${data.id}`
+
+      // modal
+      isEventType && showModal(
+        <div>
+          Cette offre est-elle soumise à des dates ou des horaires particuliers ?
+          <NavLink
+            className='button'
+            to={`${routePath}/dates`}
+          >
+            Oui
+          </NavLink>
+          <button onClick={() => { closeModal(); history.push('/offres') }}
+            className='button'>
+            Non
+          </button>
+        </div>
+      )
+    }
+  }
+
+  componentDidMount () {
+    this.handleMergeForm()
+    this.handleShowOccurencesModal()
+  }
+
+  componentDidUpdate (prevProps) {
+    const {
+      match: { params: { modalType } },
+      uniqueVenue
+    } = this.props
+    if (!prevProps.uniqueVenue && uniqueVenue) {
+      this.handleMergeForm()
+    }
+    if (!get(prevProps, 'match.params.modalType') && modalType === 'dates') {
+      this.handleShowOccurencesModal()
+    }
+  }
+
+  render () {
+    const {
+      currentOccasion,
+      isNew,
+      isReadOnly,
+      occasionIdOrNew,
+      offerForm,
+      offererOptions,
+      routePath,
+      selectedOffererId,
+      selectedVenueId,
+      venueOptions,
+      user
+    } = this.props
+    const {
+      author,
+      contactName,
+      contactEmail,
+      contactPhone,
+      description,
+      durationMinutes,
+      mediaUrls,
+      mediations,
+      performer,
+      occurences,
+      stageDirector,
+    } = (currentOccasion || {})
+    const {
+      isEventType
+    } = (offerForm || {})
+
+    const offererOptionsWithPlaceholder = get(offererOptions, 'length') > 1
+      ? [{ label: 'Sélectionnez une structure' }].concat(offererOptions)
+      : offererOptions
+
+    const venueOptionsWithPlaceholder = get(venueOptions, 'length') > 1
+      ? [{ label: 'Sélectionnez un lieu' }].concat(venueOptions)
+      : venueOptions
+
+    return (
+      <div>
+        {
+          !isNew && (
+            <div className='field'>
+            {
+              isEventType && (
+                <div className='field form-field is-horizontal'>
+                  <div className='field-label'>
+                    <label className="label" htmlFor="input_occasions_name">
+                      <div className="subtitle">Dates :</div>
+                    </label>
+                  </div>
+                  <div className='field-body'>
+                    <div className='field'>
+                      <div className='nb-dates'>
+                        {pluralize(occurences.length, 'date')}
+                      </div>
+                      <NavLink
+                        className='button is-primary is-outlined is-small'
+                        to={`${routePath}/dates`}
+                      >
+                        <span className='icon'><Icon svg='ico-calendar' /></span>
+                        <span>Gérer les dates</span>
+                      </NavLink>
+                    </div>
+                  </div>
+                </div>
+              )
+            }
+            <MediationManager
+              mediations={mediations}
+            />
+          </div>
+          )
+        }
+        <h2 className='pc-list-title'>
+          Infos pratiques
+        </h2>
+        {
+          /*
+          <FormField
+            collectionName='occasions'
+            entityId={occasionIdOrNew}
+            isHorizontal
+            label={<Label title="Prix:" />}
+            name="price"
+            readOnly={isReadOnly}
+          />
+          <FormField
+            className='column'
+            collectionName='occasions'
+            entityId={occasionIdOrNew}
+            inputClassName='input is-rounded'
+            label={<Label title="Gratuit" />}
+            name="isForFree"
+            readOnly={isReadOnly}
+            type="checkbox"
+          />
+          */
+        }
+        <FormField
+          collectionName='occasions'
+          defaultValue={selectedOffererId}
+          entityId={occasionIdOrNew}
+          isHorizontal
+          label={<Label title="Structure :" />}
+          required
+          name='offererId'
+          options={offererOptionsWithPlaceholder}
+          readOnly={isReadOnly || !isNew}
+          type="select"
+        />
+        {
+          selectedOffererId && get(venueOptions, 'length') === 0
+            ? (
+              <p>
+                Il faut obligatoirement une structure avec un lieu.
+              </p>
+            )
+            :
+              get(venueOptions, 'length') > 0 && <FormField
+                collectionName='occasions'
+                defaultValue={selectedVenueId}
+                entityId={occasionIdOrNew}
+                isHorizontal
+                label={<Label title="Lieu :" />}
+                name='venueId'
+                options={venueOptionsWithPlaceholder}
+                readOnly={isReadOnly || !isNew}
+                required={!isReadOnly}
+                type="select"
+              />
+        }
+        {
+          isEventType && (
+            <FormField
+              collectionName='occasions'
+              defaultValue={durationMinutes}
+              entityId={occasionIdOrNew}
+              isHorizontal
+              label={<Label title="Durée (en minutes) :" />}
+              name="durationMinutes"
+              readOnly={isReadOnly}
+              required={!isReadOnly}
+              type="number"
+            />
+          )
+        }
+
+        <h2 className='pc-list-title'>Infos artistiques</h2>
+        <FormField
+          collectionName='occasions'
+          defaultValue={description}
+          entityId={occasionIdOrNew}
+          isHorizontal
+          isExpanded
+          label={<Label title="Description :" />}
+          name="description"
+          readOnly={isReadOnly}
+          required={!isReadOnly}
+          type="textarea"
+        />
+        <FormField
+          collectionName='occasions'
+          defaultValue={author}
+          entityId={occasionIdOrNew}
+          isHorizontal
+          isExpanded
+          label={<Label title="Auteur :" />}
+          name="author"
+          readOnly={isReadOnly}
+        />
+        {
+          isEventType && [
+            <FormField
+              collectionName='occasions'
+              defaultValue={stageDirector}
+              entityId={occasionIdOrNew}
+              isHorizontal
+              isExpanded
+              key={0}
+              label={<Label title="Metteur en scène:" />}
+              name="stageDirector"
+              readOnly={isReadOnly}
+            />,
+            <FormField
+              collectionName='occasions'
+              defaultValue={performer}
+              entityId={occasionIdOrNew}
+              isHorizontal
+              isExpanded
+              key={1}
+              label={<Label title="Interprète:" />}
+              name="performer"
+              readOnly={isReadOnly}
+            />
+          ]
+        }
+
+        <h2 className='pc-list-title'>Contact</h2>
+        <FormField
+          collectionName='occasions'
+          defaultValue={contactName || get(user, 'publicName')}
+          entityId={occasionIdOrNew}
+          isHorizontal
+          isExpanded
+          label={<Label title="Nom du contact :" />}
+          name="contactName"
+          readOnly={isReadOnly}
+          required={!isReadOnly}
+        />
+        <FormField
+          collectionName='occasions'
+          defaultValue={contactEmail || get(user, 'email')}
+          entityId={occasionIdOrNew}
+          isHorizontal
+          isExpanded
+          label={<Label title="Email de contact :" />}
+          name="contactEmail"
+          readOnly={isReadOnly}
+          required={!isReadOnly}
+          type="email"
+        />
+        <FormField
+          collectionName='occasions'
+          defaultValue={contactPhone}
+          entityId={occasionIdOrNew}
+          isHorizontal
+          label={<Label title="Tel de contact :" />}
+          name="contactPhone"
+          readOnly={isReadOnly}
+        />
+        {false && <FormField
+                    collectionName='occasions'
+                    defaultValue={mediaUrls}
+                    entityId={occasionIdOrNew}
+                    label={<Label title="Media URLs" />}
+                    name="mediaUrls"
+                    readOnly={isReadOnly}
+                    type="list"
+                  />}
+
+
+      </div>
+    )
+  }
+}
+
+export default connect(
+  (state, ownProps) => ({
+    offererOptions: selectOffererOptions(state, ownProps),
+    selectedOffererId: selectSelectedOffererId(state, ownProps),
+    selectedVenueId: selectSelectedVenueId(state, ownProps),
+    selectedVenues: selectSelectedVenues(state, ownProps),
+    typeOptions: state.data.types,
+    venueOptions: selectVenueOptions(state, ownProps)
+  }),
+  {
+    closeModal,
+    mergeForm,
+    showModal
+  }
+)(OfferForm)
