+++ src/selectors/selectedOffererId.js
@@ -2,16 +2,14 @@ import get from 'lodash.get'
 import { createSelector } from 'reselect'
 
 import selectOfferers from './offerers'
-import { NEW } from '../utils/config'
+import selectFormOccasion from './formOccasion'
 
 export default createSelector(
   selectOfferers,
+  selectFormOccasion,
   (state, ownProps) => get(ownProps, 'currentOccasion.offererId'),
-  (state, ownProps) => get(ownProps, 'currentOccasion.id'),
-  (state, ownProps) => get(state, 'form.occasionsById'),
-  (offerers, currentOffererId, currentOccasionId, occasionsById) => {
-    const formOccasion = get(occasionsById, currentOccasionId || NEW)
-    return get(formOccasion, 'offererId') || 
+  (offerers, formOccasion, currentOffererId) => {
+    return get(formOccasion, 'offererId') ||
       currentOffererId ||
       (get(offerers, 'length') === 1 && get(offerers, '0.id'))
   }
