+++ src/selectors/selectedType.js
@@ -1,15 +1,12 @@
 import get from 'lodash.get'
 import { createSelector } from 'reselect'
 
-import { NEW } from '../utils/config'
+import selectFormOccasion from './formOccasion'
 
 export default createSelector(
-  state => state.data.types,
-  state => get(state, 'form.occasionsById'),
-  (state, ownProps) => get(ownProps, 'currentOccasion.id'),
-  (types, occasionsById, occasionId) => {
-    const typeValue = occasionsById &&
-      get(occasionsById, `${occasionId || NEW}.type`)
-    return typeValue && types && types.find(type => type.value === typeValue)
-  }
+  selectFormOccasion,
+  (state, ownProps) => get(ownProps, 'currentOccasion.typeOption'),
+  (formOccasion, typeOption) =>
+    get(formOccasion, 'type') ||
+    get(typeOption, 'value')
 )
